+++ app/controllers/recording_studio_publications/catalogue_admin_controller.rb
@@ -8,7 +8,7 @@ module RecordingStudioPublications
     before_action :set_current_actor
 
     helper_method :recording_studio_admin_context, :page_nav_anchor_url, :preserve_anchor_url,
-                  :inventory_path
+                  :inventory_path, :articles_admin_path
 
     private
 
@@ -103,6 +103,10 @@ module RecordingStudioPublications
 
     def inventory_path
       RecordingStudioPublications::Admin.publications_screen_path(recording_studio_admin_context)
+    end
+
+    def articles_admin_path(publication = nil)
+      RecordingStudioPublications::Admin.articles_screen_path(recording_studio_admin_context, publication: publication)
     end
   end
 end